+++ src/types/index.ts
@@ -1,2 +1,2 @@
 export { Settings } from './Settings';
-export { Methods } from './Methods';
+export { Callbacks } from './Callbacks';