+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 
 ## 1.0.0 beta 1 (Soon)
 #### New Feature
-- added settings: `min`, `max`, `exclude`,`hold`
+- added settings: `min`, `max`, `exclude`, `hold`
 - added methods: `onInit`, `onRandomize`, `onResult`, `onRangeEnd`